+++ backend/src/io/mod.rs
@@ -3,3 +3,4 @@ pub mod fs;
 mod hash;
 pub mod http;
 pub mod human;
+pub mod markdown;