+++ backend/src/state/block.rs
@@ -1,0 +1,7 @@
+use crate::looksyk::model::PageId;
+
+#[derive(PartialEq, Eq, Debug, Clone)]
+pub struct BlockReference {
+    pub page_id: PageId,
+    pub block_number: usize,
+}